made progress on part5

+++ part5/backendblog/blogHandlers.js
@@ -1,5 +1,5 @@
 const { verifyToken } = require("./authHandlers")
-const { getBlogs, addBlog, deleteBlog, findAndUpdateBlog } = require("./mongo")
+const { getBlogs, addBlog, deleteBlog, findAndUpdateBlog, getBlogById } = require("./mongo")
 
 
 async function getUserBlogs(request, response){
@@ -27,9 +27,7 @@ async function getUserBlogs(request, response){
    async function updateBlog(request,response){
    const jwtToken = request.token
    try{
-
       const payload = verifyToken(jwtToken)
-      console.log({payload})
       const username = payload.username
       let updatedBlog = request.body.blog
       const blogId = request.params.blogId
@@ -44,14 +42,19 @@ async function getUserBlogs(request, response){
 
   async function deleteBlogHandler(request,response){
       const user = request.username
-      const id = request.body.blogId
-      try{
-         const deletedBlog = await deleteBlog(user,id)
-         response.status(200).json(deletedBlog)
-      }catch(e){
-         response.status(500).json(e.message)
+      const id = request.params.blogId
+      const blog = await getBlogById(id)
+
+      if(blog[0].user.username != user) response.status(400).json({message:'Could not delete, not authorized'})
+      else{
+         try{
+            const deletedBlog = await deleteBlog(user,id)
+            response.status(200).json(deletedBlog)
+         }catch(e){
+            response.status(500).json(e.message)
+         }
       }
   }
 
 
-  module.exports = {getUserBlogs,createNewBlogHandler,updateBlog}
+  module.exports = {getUserBlogs,createNewBlogHandler,updateBlog,deleteBlogHandler}
